Implement stepped execution; emit debug msg
One can now choose to step through execution. This executes one cycle
and then waits for input. The idea is to implement a count step, so that
when one inputs an integer, that amount of steps will be taken before
waiting for input.

Also a message is outputted for event emissions.