Add debugging to the simulator
After quite a struggle with Tkinter (Python's Tk library) and piping to
stdin, it finally works. Every state machine created starts a small
window containing only a textfield. When text is written to the
window's process' stdin, it is used in the textfield.

This has also been implemented in the simulator, so a cycle shows
information in a machine's complementary window.