Add numbers to events for debuggin
Every new event gets an incremented number, which is displayed in the
debug windows. This makes it much easier to trace back a program's
execution.

Also a reset method has been added to MachineControl, which should
prepare machine control for a next round of execution. This currently
only resets the event counter, the rest is TODO.